refactored path to user-list component
