+++ Comp354Project/Project-UI/src/app/app-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { UserListComponent } from './user-list/user-list.component';
+import { UserListComponent } from './Components/user-list/user-list.component';
 
 const routes: Routes = [
   { path: 'users', component: UserListComponent }
