+++ Comp354Project/Project-UI/src/app/app.module.ts
@@ -4,7 +4,7 @@ import { AppRoutingModule } from './app-routing.module';
 import { FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
 import { AppComponent } from './app.component';
-import { UserListComponent } from './user-list/user-list.component';
+import { UserListComponent } from './Components/user-list/user-list.component';
 
 
 import { UserService } from './service/user.service';
